+++ bin/encodePassword.php
@@ -5,7 +5,7 @@ use Silex\Provider\SecurityServiceProvider;
 
 require __DIR__.'/../vendor/autoload.php';
 
-$configPath = __DIR__.'/../config.php';
+$configPath = __DIR__.'/../config/config.php';
 
 $app = new Application(
     file_exists($configPath) ? require $configPath : []