+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - command line boilerplate to home page.
+- site favicon, image from [Bytesize](https://danklammer.com/bytesize-icons/).
 
 ### Changed
 - home page markup to Bootstrap 4.