+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - command line boilerplate to home page.
 - site favicon, image from [Bytesize](https://danklammer.com/bytesize-icons/).
+- default error log configuration detection from PHP settings.
 
 ### Changed
 - home page markup to Bootstrap 4.